# aspnetcore-server-generated/src/IO.Swagger/Controllers/DefaultApi.cs
/*
 * Test Task for Akvelon by Andrei Kekov
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DefaultApiController : ControllerBase
    { 
        /// <summary>
        /// Add a new project
        /// </summary>
        /// <param name="body">New project data</param>
        /// <response code="200">Successful operation</response>
        /// <response code="405">Invalid input</response>
        [HttpPost]
        [Route("/project")]
        [ValidateModelState]
        [SwaggerOperation("AddProject")]
        [SwaggerResponse(statusCode: 200, type: typeof(Project), description: "Successful operation")]
        public virtual IActionResult AddProject([FromBody]Project body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Project));

            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405);
            string exampleJson = null;
            exampleJson = "{\n  \"statusId\" : 2,\n  \"name\" : \"Dinner\",\n  \"completionDate\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"id\" : 42,\n  \"priority\" : 200,\n  \"startDate\" : \"2000-01-23T04:56:07.000+00:00\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Project>(exampleJson)
                        : default(Project);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Add a new task to a project
        /// </summary>
        /// <param name="body">New task data</param>
        /// <response code="200">Successful operation</response>
        /// <response code="404">Project ID not found</response>
        /// <response code="405">Invalid input</response>
        [HttpPost]
        [Route("/task")]
        [ValidateModelState]
        [SwaggerOperation("AddTask")]
        [SwaggerResponse(statusCode: 200, type: typeof(Task), description: "Successful operation")]
        public virtual IActionResult AddTask([FromBody]Task body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Task));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405);
            string exampleJson = null;
            exampleJson = "{\n  \"statusId\" : 2,\n  \"name\" : \"Cut the potatoes\",\n  \"description\" : \"Peel the potatoes, then cut them into slices\",\n  \"id\" : 65536,\n  \"priority\" : 200,\n  \"projectId\" : 42\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Task>(exampleJson)
                        : default(Task);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete the project with {projectId}
        /// </summary>
        /// <param name="projectId">ID of project to delete</param>
        /// <response code="200">Successful operation</response>
        /// <response code="404">Project ID not found</response>
        [HttpDelete]
        [Route("/project/{projectId}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteProject")]
        public virtual IActionResult DeleteProject([FromRoute][Required]int? projectId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            throw new NotImplementedException();
        }

        /// <summary>
        /// View all existing projects
        /// </summary>
        /// <response code="200">Successful operation</response>
        [HttpGet]
        [Route("/project")]
        [ValidateModelState]
        [SwaggerOperation("GetProjects")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Project>), description: "Successful operation")]
        public virtual IActionResult GetProjects()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Project>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"statusId\" : 2,\n  \"name\" : \"Dinner\",\n  \"completionDate\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"id\" : 42,\n  \"priority\" : 200,\n  \"startDate\" : \"2000-01-23T04:56:07.000+00:00\"\n}, {\n  \"statusId\" : 2,\n  \"name\" : \"Dinner\",\n  \"completionDate\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"id\" : 42,\n  \"priority\" : 200,\n  \"startDate\" : \"2000-01-23T04:56:07.000+00:00\"\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Project>>(exampleJson)
                        : default(List<Project>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// View all tasks for a project
        /// </summary>
        /// <param name="projectId">ID of project to view tasks</param>
        /// <response code="200">Successful operation</response>
        /// <response code="404">Project ID not found</response>
        [HttpGet]
        [Route("/project/{projectId}/tasks")]
        [ValidateModelState]
        [SwaggerOperation("GetTasks")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Task>), description: "Successful operation")]
        public virtual IActionResult GetTasks([FromRoute][Required]int? projectId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Task>));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "[ {\n  \"statusId\" : 2,\n  \"name\" : \"Cut the potatoes\",\n  \"description\" : \"Peel the potatoes, then cut them into slices\",\n  \"id\" : 65536,\n  \"priority\" : 200,\n  \"projectId\" : 42\n}, {\n  \"statusId\" : 2,\n  \"name\" : \"Cut the potatoes\",\n  \"description\" : \"Peel the potatoes, then cut them into slices\",\n  \"id\" : 65536,\n  \"priority\" : 200,\n  \"projectId\" : 42\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Task>>(exampleJson)
                        : default(List<Task>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete the task with {taskId}
        /// </summary>
        /// <param name="taskId">ID of task to delete</param>
        /// <response code="200">Successful operation</response>
        /// <response code="404">Task ID not found</response>
        [HttpDelete]
        [Route("/task/{taskId}")]
        [ValidateModelState]
        [SwaggerOperation("TaskTaskIdDelete")]
        public virtual IActionResult TaskTaskIdDelete([FromRoute][Required]int? taskId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Update an existing project
        /// </summary>
        /// <param name="body">New project data</param>
        /// <response code="200">Successful operation</response>
        /// <response code="404">Project ID not found</response>
        /// <response code="405">Invalid input</response>
        [HttpPut]
        [Route("/project")]
        [ValidateModelState]
        [SwaggerOperation("UpdateProject")]
        public virtual IActionResult UpdateProject([FromBody]Project body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Update an existing task
        /// </summary>
        /// <param name="body">New task data</param>
        /// <response code="200">Successful operation</response>
        /// <response code="404">Project ID or Task ID not found</response>
        /// <response code="405">Invalid input</response>
        [HttpPut]
        [Route("/task")]
        [ValidateModelState]
        [SwaggerOperation("UpdateTask")]
        public virtual IActionResult UpdateTask([FromBody]Task body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405);

            throw new NotImplementedException();
        }
    }
}